+++ src/SMAPI.Toolkit/Framework/ModData/ModWarning.cs
@@ -19,6 +19,7 @@ namespace StardewModdingAPI.Toolkit.Framework.ModData
         PatchesGame = 4,
 
         /// <summary>The mod uses the <c>dynamic</c> keyword which won't work on Linux/macOS.</summary>
+        [Obsolete("This value is no longer used by SMAPI and will be removed in the upcoming SMAPI 4.0.0.")]
         UsesDynamic = 8,
 
         /// <summary>The mod references specialized 'unvalidated update tick' events which may impact stability.</summary>