+++ src/SMAPI/Framework/ModHelpers/ModHelper.cs
@@ -33,7 +33,7 @@ namespace StardewModdingAPI.Framework.ModHelpers
             get
             {
                 SCore.DeprecationManager.Warn(
-                    source: SCore.DeprecationManager.GetMod(this.ModID),
+                    source: this.Mod,
                     nounPhrase: $"{nameof(IModHelper)}.{nameof(IModHelper.Content)}",
                     version: "3.14.0",
                     severity: DeprecationLevel.Notice